add loading screen with gif for inital page load
