+++ src/components/Loading/Loading.js
@@ -1,15 +1,30 @@
 import React, { Component } from 'react';
-import { Button, Image, View, Text } from 'react-native';
+import { Button, Image, View, Text, StyleSheet } from 'react-native';
 
 class Loading extends Component {
   render() {
+
     return (
-      <Image
-        source={require('../../images/BarStock_Logo.png')}
-        style={{width: 150, height: 33}}
-      />
+      <View style={styles.container}>
+        <Image source={require('../../../assets/loading.gif')} />
+      </View>
     );
   }
 }
+
+const styles = StyleSheet.create({
+  container: {
+    backgroundColor: '#ffffff',
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center"
+  },
+
+  loadingText: {
+    color: 'white',
+    fontSize: 40,
+    fontWeight: 'bold'
+  },
+})
 
 export default Loading
